changed dtype

+++ DAOT_mnist/main.py
@@ -65,7 +65,7 @@ def loss_fn(model, features, config, training):
         model.trainable_variables if 'bias' not in v.name])
 
     model_output = model(inputs, training=training)
-    model_output = tf.cast(model_output, dtype = tf.int32)
+    model_output = tf.cast(model_output, dtype = tf.float32)
     print(model_output)
     classification_loss = tf.losses.binary_crossentropy(
         tf.one_hot(label, axis=-1, depth=config.num_classes),
